Package restructure
This removes a couple of unnecessary files from GitHub and npm and ensures that the remaining files have a unique location each without having to copy things around.

Also introduces the docs/ directory to be able to serve up to date versions on GitHub pages while also loading the common compiled JS version from that location.